Fix user model usage
